adds additional tests

+++ src/test/java/io/supertokens/test/userRoles/UserRolesStorageTest.java
@@ -28,6 +28,7 @@ import io.supertokens.test.TestingProcessManager;
 import io.supertokens.test.Utils;
 import io.supertokens.userroles.UserRoles;
 import io.supertokens.version.Version;
+import jdk.jshell.execution.Util;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.Rule;
@@ -771,6 +772,33 @@ public class UserRolesStorageTest {
 
             boolean didRoleExist = storage.deleteRole(role);
             assertFalse(didRoleExist);
+        }
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void testGettingAllCreatedRoles() throws Exception {
+        String[] args = { "../" };
+        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        if (StorageLayer.getStorage(process.getProcess()).getType() != STORAGE_TYPE.SQL) {
+            return;
+        }
+        UserRolesSQLStorage storage = StorageLayer.getUserRolesStorage(process.main);
+
+        // create roles
+        String[] roles = new String[] { "role1", "role2", "role3" };
+        for (String role : roles) {
+            UserRoles.createNewRoleOrModifyItsPermissions(process.main, role, null);
+        }
+
+        // retrieve all role and check for correct output
+        {
+            String[] retrievedRoles = storage.getRoles();
+            Utils.checkThatArraysAreEqual(roles, retrievedRoles);
         }
 
         process.kill();
